src/requirements.txt
wheel
setuptools
PyVCF
blist
